Half language change issue on back gesture solved.
Half language change issue on back gesture ios solved.
